Split out Template interface to own file
